# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RubinTV",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/locations/{location}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Location Detail",
        "operationId": "get_location_detail_api_locations__location__get",
        "parameters": [
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/locations/{location}/cameras/{camera}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Camera Detail",
        "operationId": "get_camera_detail_api_locations__location__cameras__camera__get",
        "parameters": [
          {
            "name": "camera",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera"
            }
          },
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/locations/{location}/cameras/{camera}/calendar": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Calendar",
        "operationId": "get_calendar_api_locations__location__cameras__camera__calendar_get",
        "parameters": [
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location"
            }
          },
          {
            "name": "camera",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/locations/{location}/cameras/{camera}/dates/{date}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Date Payload",
        "operationId": "get_date_payload_api_locations__location__cameras__camera__dates__date__get",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          },
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location"
            }
          },
          {
            "name": "camera",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatePayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/locations/{location}/cameras/{camera}/metadata/{date}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Metadata",
        "operationId": "get_metadata_api_locations__location__cameras__camera__metadata__date__get",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          },
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location"
            }
          },
          {
            "name": "camera",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Metadata Api Locations  Location  Cameras  Camera  Metadata  Date  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/locations/{location}/cameras/{camera}/events": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Event By Key",
        "description": "Resolve a specific event by its S3 key (validated by the parser).",
        "operationId": "get_event_by_key_api_locations__location__cameras__camera__events_get",
        "parameters": [
          {
            "name": "camera",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera"
            }
          },
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location"
            }
          },
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/locations/{location}/cameras/{camera}/night-report/{date}": {
      "get": {
        "tags": [
          "night-report"
        ],
        "summary": "Get Night Report",
        "operationId": "get_night_report_api_locations__location__cameras__camera__night_report__date__get",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          },
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location"
            }
          },
          {
            "name": "camera",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NightReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/locations/{location}/admin/controls": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Controls",
        "operationId": "get_controls_api_locations__location__admin_controls_get",
        "parameters": [
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Set Control",
        "operationId": "set_control_api_locations__location__admin_controls_post",
        "parameters": [
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location"
            }
          },
          {
            "name": "x-auth-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlValue"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlValue"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CalendarOut": {
        "properties": {
          "dates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dates"
          }
        },
        "type": "object",
        "required": [
          "dates"
        ],
        "title": "CalendarOut"
      },
      "CameraGroupOut": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "cameras": {
            "items": {
              "$ref": "#/components/schemas/CameraSummary"
            },
            "type": "array",
            "title": "Cameras"
          }
        },
        "type": "object",
        "required": [
          "label",
          "cameras"
        ],
        "title": "CameraGroupOut"
      },
      "CameraOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "online": {
            "type": "boolean",
            "title": "Online"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelOut"
            },
            "type": "array",
            "title": "Channels"
          },
          "metadata_columns": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Metadata Columns"
          },
          "image_viewer_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Viewer Link"
          },
          "has_mosaic": {
            "type": "boolean",
            "title": "Has Mosaic"
          },
          "has_allsky": {
            "type": "boolean",
            "title": "Has Allsky"
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "online",
          "channels",
          "metadata_columns",
          "image_viewer_link",
          "has_mosaic",
          "has_allsky"
        ],
        "title": "CameraOut",
        "description": "Full camera detail for the camera page header."
      },
      "CameraSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "online": {
            "type": "boolean",
            "title": "Online"
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "online"
        ],
        "title": "CameraSummary",
        "description": "Camera as listed under a location (no heavy detail)."
      },
      "ChannelOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "colour": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colour"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "per_day": {
            "type": "boolean",
            "title": "Per Day"
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "label",
          "colour",
          "icon",
          "per_day"
        ],
        "title": "ChannelOut"
      },
      "ControlValue": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "ControlValue"
      },
      "ControlsOut": {
        "properties": {
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "values"
        ],
        "title": "ControlsOut"
      },
      "DatePayload": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "channels": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "string"
                  }
                ]
              },
              "type": "array"
            },
            "type": "object",
            "title": "Channels"
          },
          "extensions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ExtInfoOut"
            },
            "type": "object",
            "title": "Extensions"
          },
          "per_day": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Per Day"
          },
          "metadata": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Metadata"
          },
          "has_night_report": {
            "type": "boolean",
            "title": "Has Night Report"
          }
        },
        "type": "object",
        "required": [
          "date",
          "channels",
          "extensions",
          "per_day",
          "metadata",
          "has_night_report"
        ],
        "title": "DatePayload",
        "description": "Everything needed to render a camera table for one date."
      },
      "EventOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "camera": {
            "type": "string",
            "title": "Camera"
          },
          "day_obs": {
            "type": "string",
            "title": "Day Obs"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "seq_num": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Seq Num"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "ext": {
            "type": "string",
            "title": "Ext"
          }
        },
        "type": "object",
        "required": [
          "key",
          "camera",
          "day_obs",
          "channel",
          "seq_num",
          "filename",
          "ext"
        ],
        "title": "EventOut"
      },
      "ExtInfoOut": {
        "properties": {
          "default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "exceptions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Exceptions"
          }
        },
        "type": "object",
        "required": [
          "default",
          "exceptions"
        ],
        "title": "ExtInfoOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "camera_groups": {
            "items": {
              "$ref": "#/components/schemas/CameraGroupOut"
            },
            "type": "array",
            "title": "Camera Groups"
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "camera_groups"
        ],
        "title": "LocationOut"
      },
...
      "NightReportOut": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "exists": {
            "type": "boolean",
            "title": "Exists"
          },
          "text": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Text"
          },
          "plots": {
            "items": {
              "$ref": "#/components/schemas/PlotOut"
            },
            "type": "array",
            "title": "Plots"
          }
        },
        "type": "object",
        "required": [
          "date",
          "exists",
          "text",
          "plots"
        ],
        "title": "NightReportOut"
      },
      "PlotOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "key",
          "group",
          "filename"
        ],
        "title": "PlotOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}